requirements.txt
flask-cors
requests
openai
mysql-connector-python
python-dotenv
bcrypt